Fix bifurcated gameClient issue
Prior, one gameClient is created & passed to the board, while another is
used to make automatic moves. This meant that when a move was made
_after_ automatic moves, the new move would be made on a gameClient with
no moves, and would reset the board.
This fix stores & updates the gameClientRef in state, so that Board is
always up to date with the latest gameClient.